requirements.txt
opencv-python==4.9.0.80
numpy==1.26.4
pandas==2.2.2
tensorflow==2.15.0
tqdm==4.66.2
requests==2.31.0
paho-mqtt==1.6.1
pymongo==4.6.3
scikit-learn==1.4.2
